Part I: RecursiveArrayTools as weak dep.
