+++ src/groups/product_group.jl
@@ -52,10 +52,6 @@ function adjoint_inv_diff!(G::ProductGroup, Y, p, X)
     return Y
 end
 
-function identity_element(G::ProductGroup)
-    M = G.manifold
-    return ArrayPartition(map(identity_element, M.manifolds))
-end
 function identity_element!(G::ProductGroup, p)
     pes = submanifold_components(G, p)
     M = G.manifold
@@ -127,16 +123,6 @@ function inv_diff!(G::ProductGroup, Y, p, X)
 end
 
 _compose(G::ProductGroup, p, q) = _compose(G.manifold, p, q)
-function _compose(M::ProductManifold, p::ArrayPartition, q::ArrayPartition)
-    return ArrayPartition(
-        map(
-            compose,
-            M.manifolds,
-            submanifold_components(M, p),
-            submanifold_components(M, q),
-        )...,
-    )
-end
 
 _compose!(G::ProductGroup, x, p, q) = _compose!(G.manifold, x, p, q)
 function _compose!(M::ProductManifold, x, p, q)
@@ -148,23 +134,6 @@ function _compose!(M::ProductManifold, x, p, q)
         submanifold_components(M, q),
     )
     return x
-end
-
-function translate(
-    M::ProductGroup,
-    p::ArrayPartition,
-    q::ArrayPartition,
-    conv::ActionDirectionAndSide,
-)
-    return ArrayPartition(
-        map(
-            translate,
-            M.manifold.manifolds,
-            submanifold_components(M, p),
-            submanifold_components(M, q),
-            repeated(conv),
-        )...,
-    )
 end
 
 function translate!(M::ProductGroup, x, p, q, conv::ActionDirectionAndSide)
@@ -179,19 +148,6 @@ function translate!(M::ProductGroup, x, p, q, conv::ActionDirectionAndSide)
     return x
 end
 
-function inverse_translate(G::ProductGroup, p, q, conv::ActionDirectionAndSide)
-    M = G.manifold
-    return ArrayPartition(
-        map(
-            inverse_translate,
-            M.manifolds,
-            submanifold_components(G, p),
-            submanifold_components(G, q),
-            repeated(conv),
-        )...,
-    )
-end
-
 function inverse_translate!(G::ProductGroup, x, p, q, conv::ActionDirectionAndSide)
     M = G.manifold
     map(
@@ -203,20 +159,6 @@ function inverse_translate!(G::ProductGroup, x, p, q, conv::ActionDirectionAndSi
         repeated(conv),
     )
     return x
-end
-
-function translate_diff(G::ProductGroup, p, q, X, conv::ActionDirectionAndSide)
-    M = G.manifold
-    return ArrayPartition(
-        map(
-            translate_diff,
-            M.manifolds,
-            submanifold_components(G, p),
-            submanifold_components(G, q),
-            submanifold_components(G, X),
-            repeated(conv),
-        )...,
-    )
 end
 
 function translate_diff!(G::ProductGroup, Y, p, q, X, conv::ActionDirectionAndSide)
@@ -233,19 +175,7 @@ function translate_diff!(G::ProductGroup, Y, p, q, X, conv::ActionDirectionAndSi
     return Y
 end
 
-function inverse_translate_diff(G::ProductGroup, p, q, X, conv::ActionDirectionAndSide)
-    M = G.manifold
-    return ArrayPartition(
-        map(
-            inverse_translate_diff,
-            M.manifolds,
-            submanifold_components(G, p),
-            submanifold_components(G, q),
-            submanifold_components(G, X),
-            repeated(conv),
-        )...,
-    )
-end
+
 
 function inverse_translate_diff!(G::ProductGroup, Y, p, q, X, conv::ActionDirectionAndSide)
     M = G.manifold
@@ -261,16 +191,7 @@ function inverse_translate_diff!(G::ProductGroup, Y, p, q, X, conv::ActionDirect
     return Y
 end
 
-function Base.exp(M::ProductGroup, p::Identity{ProductOperation}, X::ArrayPartition)
-    return ArrayPartition(
-        map(
-            exp,
-            M.manifold.manifolds,
-            submanifold_components(M, p),
-            submanifold_components(M, X),
-        )...,
-    )
-end
+
 
 function exp!(M::ProductGroup, q, p::Identity{ProductOperation}, X)
     map(
@@ -283,26 +204,10 @@ function exp!(M::ProductGroup, q, p::Identity{ProductOperation}, X)
     return q
 end
 
-function exp_lie(G::ProductGroup, X)
-    M = G.manifold
-    return ArrayPartition(map(exp_lie, M.manifolds, submanifold_components(G, X))...)
-end
-
 function exp_lie!(G::ProductGroup, q, X)
     M = G.manifold
     map(exp_lie!, M.manifolds, submanifold_components(G, q), submanifold_components(G, X))
     return q
-end
-
-function Base.log(M::ProductGroup, p::Identity{ProductOperation}, q::ArrayPartition)
-    return ArrayPartition(
-        map(
-            log,
-            M.manifold.manifolds,
-            submanifold_components(M, p),
-            submanifold_components(M, q),
-        )...,
-    )
 end
 
 function log!(M::ProductGroup, X, p::Identity{ProductOperation}, q)
@@ -330,36 +235,8 @@ function log_lie!(G::ProductGroup, X, q::Identity{ProductOperation})
     return X
 end
 
-Base.@propagate_inbounds function Base.getindex(
-    p::ArrayPartition,
-    M::ProductGroup,
-    i::Union{Integer,Colon,AbstractVector,Val},
-)
-    return getindex(p, base_manifold(M), i)
-end
-
-Base.@propagate_inbounds function Base.setindex!(
-    q::ArrayPartition,
-    p,
-    M::ProductGroup,
-    i::Union{Integer,Colon,AbstractVector,Val},
-)
-    return setindex!(q, p, base_manifold(M), i)
-end
 
 # these isapprox methods are here just to reduce time-to-first-isapprox
-function isapprox(G::ProductGroup, p::ArrayPartition, q::ArrayPartition; kwargs...)
-    return isapprox(G.manifold, p, q; kwargs...)
-end
-function isapprox(
-    G::ProductGroup,
-    p::ArrayPartition,
-    X::ArrayPartition,
-    Y::ArrayPartition;
-    kwargs...,
-)
-    return isapprox(G.manifold, p, X, Y; kwargs...)
-end
 function isapprox(G::ProductGroup, ::Identity{ProductOperation}, X, Y; kwargs...)
     return isapprox(G.manifold, identity_element(G), X, Y; kwargs...)
 end
