+++ src/manifolds/ProductManifold.jl
@@ -36,26 +36,7 @@ struct ProductPointDistribution{
     distributions::TD
 end
 
-function adjoint_Jacobi_field(
-    M::ProductManifold,
-    p::ArrayPartition,
-    q::ArrayPartition,
-    t,
-    X::ArrayPartition,
-    β::Tβ,
-) where {Tβ}
-    return ArrayPartition(
-        map(
-            adjoint_Jacobi_field,
-            M.manifolds,
-            submanifold_components(M, p),
-            submanifold_components(M, q),
-            ntuple(_ -> t, length(M.manifolds)),
-            submanifold_components(M, X),
-            ntuple(_ -> β, length(M.manifolds)),
-        )...,
-    )
-end
+
 function adjoint_Jacobi_field!(M::ProductManifold, Y, p, q, t, X, β::Tβ) where {Tβ}
     map(
         adjoint_Jacobi_field!,
@@ -80,26 +61,6 @@ to the corresponding entry in `p`) separately.
 """
 flat(::ProductManifold, ::Any...)
 
-function jacobi_field(
-    M::ProductManifold,
-    p::ArrayPartition,
-    q::ArrayPartition,
-    t,
-    X::ArrayPartition,
-    β::Tβ,
-) where {Tβ}
-    return ArrayPartition(
-        map(
-            jacobi_field,
-            M.manifolds,
-            submanifold_components(M, p),
-            submanifold_components(M, q),
-            ntuple(_ -> t, length(M.manifolds)),
-            submanifold_components(M, X),
-            ntuple(_ -> β, length(M.manifolds)),
-        )...,
-    )
-end
 function jacobi_field!(M::ProductManifold, Y, p, q, t, X, β::Tβ) where {Tβ}
     map(
         jacobi_field!,
@@ -122,19 +83,6 @@ Return the volume of [`ProductManifold`](https://juliamanifolds.github.io/Manifo
 """
 manifold_volume(M::ProductManifold) = mapreduce(manifold_volume, *, M.manifolds)
 
-function ProductFVectorDistribution(distributions::FVectorDistribution...)
-    M = ProductManifold(map(d -> support(d).space.manifold, distributions)...)
-    fiber_type = support(distributions[1]).space.fiber_type
-    if !all(d -> support(d).space.fiber_type == fiber_type, distributions)
-        error(
-            "Not all distributions have support in vector spaces of the same type, which is currently not supported",
-        )
-    end
-    # Probably worth considering sum spaces in the future?
-    p = ArrayPartition(map(d -> support(d).space.point, distributions)...)
-    return ProductFVectorDistribution(Fiber(M, p, fiber_type), distributions)
-end
-
 function ProductPointDistribution(M::ProductManifold, distributions::MPointDistribution...)
     return ProductPointDistribution{typeof(M),typeof(distributions)}(M, distributions)
 end
@@ -143,12 +91,6 @@ function ProductPointDistribution(distributions::MPointDistribution...)
     return ProductPointDistribution(M, distributions...)
 end
 
-function Random.rand(rng::AbstractRNG, d::ProductPointDistribution)
-    return ArrayPartition(map(d -> rand(rng, d), d.distributions)...)
-end
-function Random.rand(rng::AbstractRNG, d::ProductFVectorDistribution)
-    return ArrayPartition(map(d -> rand(rng, d), d.distributions)...)
-end
 
 function Distributions._rand!(
     rng::AbstractRNG,
@@ -157,18 +99,7 @@ function Distributions._rand!(
 )
     return copyto!(x, rand(rng, d))
 end
-function Distributions._rand!(
-    rng::AbstractRNG,
-    d::ProductPointDistribution,
-    p::ArrayPartition,
-)
-    map(
-        (t1, t2) -> Distributions._rand!(rng, t1, t2),
-        d.distributions,
-        submanifold_components(d.manifold, p),
-    )
-    return p
-end
+
 function Distributions._rand!(
     rng::AbstractRNG,
     d::ProductFVectorDistribution,
@@ -176,18 +107,7 @@ function Distributions._rand!(
 )
     return copyto!(v, rand(rng, d))
 end
-function Distributions._rand!(
-    rng::AbstractRNG,
-    d::ProductFVectorDistribution,
-    X::ArrayPartition,
-)
-    map(
-        t -> Distributions._rand!(rng, t[1], t[2]),
-        d.distributions,
-        submanifold_components(d.space.manifold, X),
-    )
-    return X
-end
+
 
 @doc raw"""
     Y = riemannian_Hessian(M::ProductManifold, p, G, H, X)
